chores: extract shared from each api to utils
