+++ goldenmaster/tb_empty_impl/empty_interface.py
@@ -1,5 +1,5 @@
 from tb_empty_api import api
-from tb_empty_api.shared import EventHook
+from utils.eventhook import EventHook
 from typing import Iterable
 
 class EmptyInterface(api.IEmptyInterface):
