+++ goldenmaster/tb_same1_olink/sinks.py
@@ -2,7 +2,7 @@ import asyncio
 from typing import Any
 from olink.core import Name
 from olink.client import IObjectSink, ClientNode
-from tb_same1_api.shared import EventHook
+from utils.eventhook import EventHook
 from tb_same1_api import api
 import logging
