+++ goldenmaster/tb_same1_olink/sources.py
@@ -1,7 +1,7 @@
 from olink.core import Name
 from olink.remote import IObjectSource, RemoteNode
 from tb_same1_api import api
-from tb_same1_api.shared import EventHook
+from utils.eventhook import EventHook
 from typing import Any
 import logging
 class SameStruct1InterfaceSource(IObjectSource):
