+++ goldenmaster/tb_same2_impl/same_enum2_interface.py
@@ -1,5 +1,5 @@
 from tb_same2_api import api
-from tb_same2_api.shared import EventHook
+from utils.eventhook import EventHook
 from typing import Iterable
 
 class SameEnum2Interface(api.ISameEnum2Interface):
