+++ goldenmaster/utils/eventhook.py
