--- templates/api/shared.py
@@ -1,16 +1,0 @@
-class EventHook(object):
-    def __init__(self):
-        self.__handlers = []
-
-    def __iadd__(self, handler):
-        self.__handlers.append(handler)
-        return self
-
-    def __isub__(self, handler):
-        self.__handlers.remove(handler)
-        return self
-
-    def fire(self, *args, **keywargs):
-        for handler in self.__handlers:
-            handler(*args, **keywargs)
-
